LayoutTests/http/tests/usercontentfilter/character-set-basic-support-expected.txt

Test basic cases of filter using a character set (e.g. [a-z]).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


URL: http://127.0.0.1:8000/resources/square100.png is not blocked.
PASS isBlocked is false
URL: http://127.0.0.1:8000/resources/square100.png?casesEnsitive is not blocked.
PASS isBlocked is false
URL: http://127.0.0.1:8000/resources/square100.png?casesenSitive is not blocked.
PASS isBlocked is false
URL: http://127.0.0.1:8000/resources/square100.png?BaSiC-TeSt-cAsE is blocked.
PASS isBlocked is true
URL: http://127.0.0.1:8000/resources/square100.png?any-scheme-matcher is blocked.
PASS isBlocked is true
URL: http://127.0.0.1:8000/resources/square100.png?casesensitive is blocked.
PASS isBlocked is true
URL: http://127.0.0.1:8000/resources/square100.png?caseSeNsitive is blocked.
PASS isBlocked is true
PASS successfullyParsed is true

TEST COMPLETE
